+++ Experimental/DedicatedGameServer/Assets/Scripts/Runtime/Game/Characters/ICharacter.cs
@@ -1,0 +1,9 @@
+using Unity.Netcode;
+
+namespace Unity.DedicatedGameServerSample.Runtime
+{
+    internal interface ICharacter
+    {
+        NetworkObject NetworkObject { get; }
+    }
+}