+++ Experimental/DedicatedGameServer/Assets/Scripts/Runtime/Game/GameEvents.cs
@@ -1,0 +1,14 @@
+namespace Unity.DedicatedGameServerSample.Runtime
+{
+    internal class ResumeButtonClickedEvent : AppEvent { }
+
+    internal class QuitButtonClickedEvent : AppEvent { }
+
+    internal class MatchEndAcknowledgedEvent : AppEvent { }
+
+    internal class StartMatchEvent : AppEvent { }
+
+    internal class MenuToggleEvent : AppEvent { }
+
+    internal class EndMatchEvent : AppEvent { }
+}