+++ Experimental/DedicatedGameServer/Assets/Tests/Runtime/TestsApplication/TestsView.cs
@@ -1,0 +1,6 @@
+using Unity.DedicatedGameServerSample.Runtime;
+
+namespace Unity.DedicatedGameServerSample.Tests.Runtime
+{
+    class TestsView : View<TestsApplication> { }
+}